split up tests into modules
